+++ src/contexts/AuthProvider.tsx
@@ -8,6 +8,7 @@ import { useAuthDependencies } from "@/hooks/auth/init/useAuthDependencies";
 import { applySsrAuthState } from "@/lib/auth/init/applySsrAuthState";
 import { useAuthActions } from "@/hooks/auth/actions";
 import { useAuthSideEffects } from "@/hooks/auth/sideEffects";
+import { useStaleSessionRecovery } from "@/hooks/auth/sideEffects/useStaleSessionRecovery";
 import { useAuthValue } from "@/hooks/auth/init/useAuthValue";
 import { useLanguageSync } from "@/hooks/useLanguageSync";
 import { logger } from "@/lib/logging";
@@ -66,42 +67,7 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({
     return data?.currentUser?.user ?? null;
   }, [refetch]);
 
-  // auth:token-expired イベントをキャッチしてセッションをクリアし、LIFF 再認証を行う
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    let isHandling = false;
-
-    const handleTokenExpired = async (event: Event) => {
-      if (isHandling) return;
-      isHandling = true;
-      try {
-        logger.warn("[AuthProvider] Stale session detected — clearing and re-authenticating", {
-          detail: (event as CustomEvent).detail,
-        });
-        try {
-          await fetch("/api/sessionLogout", { method: "POST" });
-        } catch (e) {
-          logger.warn("[AuthProvider] Failed to clear session cookie", { error: e });
-        }
-        if (authStateManager) {
-          hasInitialized.current = false;
-          await initAuth({
-            communityConfig,
-            liffService,
-            authStateManager,
-            ssrCurrentUser: null,
-            ssrLineAuthenticated: false,
-            ssrPhoneAuthenticated: false,
-          });
-        }
-      } finally {
-        isHandling = false;
-      }
-    };
-
-    window.addEventListener("auth:token-expired", handleTokenExpired);
-    return () => window.removeEventListener("auth:token-expired", handleTokenExpired);
-  }, [authStateManager, communityConfig, liffService]);
+  useStaleSessionRecovery({ authStateManager, liffService, communityConfig, hasInitialized });
 
   useAuthSideEffects({ authStateManager, liffService, refetchUser, hasFullAuth });
